map routing function sample
remake map route function to make it easier to understand and add new nodes for expansion. a new functionality was added to find new route when user press an icon on the map (just working for cafeteria by now), however after implementing this function, the map will zoom after spawining the closes road segment to the user (in this case the position of the user is still fixed to ingeniería's building), pending bug to fix.
push made to be able to make pulls